requirements-txt.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
python-dotenv==0.19.0
pinecone-client==2.2.1
transformers==4.11.3
torch==1.9.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1  # If using PostgreSQL
beautifulsoup4==4.9.3
